Remove laser pointer demo, put controller rendering code in Tools, to be deleted later.
